TEST-892 - Error msg for Wrong file selection,drag & drop
