+++ src/features/provider/upload/slice.ts
@@ -20,8 +20,8 @@
 
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
+import { Status } from '../../../enums';
 import { ProcessReport } from '../../../models/ProcessReport';
-import { Status } from '../../../utils/constants';
 import { IUploadFileSlice } from './types';
 
 const initialState: IUploadFileSlice = {
